+++ cpp/unittest/server/util_test.cpp
@@ -19,7 +19,7 @@ using namespace zilliz::milvus;
 
 namespace {
 
-static const std::string LOG_FILE_PATH = "./milvus/conf/log_config.conf";
+static const char* LOG_FILE_PATH = "./milvus/conf/log_config.conf";
 
 }